Extract deployment related tasks to remote_source deployment module
